==> part3.py <==
import pandas as pd

def get_filtered_target_grouped():
    part_2_results = pd.read_csv('results/part2/part_2_result_with_dssp_location.csv')

    # create new column with boolean values
    part_2_results['in_list'] = part_2_results.apply(
        lambda row: row['amino_acid_original'] in row['list_amino_acid_variants'], axis=1)

    # filter DataFrame based on boolean column
    filtered_df = part_2_results[part_2_results['in_list']]
    filtered_df = filtered_df.drop('Unnamed: 0', axis=1)
    filtered_df = filtered_df.drop('in_list', axis=1)
    filtered_df = filtered_df.drop('gene', axis=1)
    filtered_df = filtered_df.drop('location', axis=1)
    filtered_df = filtered_df.drop('description', axis=1)
    filtered_df = filtered_df.drop('antigen_name', axis=1)
    filtered_df = filtered_df.drop('antigen_chain', axis=1)
    filtered_df = filtered_df.drop('antigen_species', axis=1)
    filtered_df.to_csv('filtered_df.csv')

    filtered_df = pd.read_csv('results/part2/part_2_result_test_for_part3.csv')

    filtered_df = filtered_df.groupby('pdb_id')['target'].apply(lambda x: ''.join(x)).reset_index()
    filtered_df.to_csv('filtered_df_grouped.csv')

    return filtered_df

def main():
    filtered_df = get_filtered_target_grouped()



    amino_acid_name_to_abbreviation = {'CYS': 'C', 'ASP': 'D', 'SER': 'S', 'GLN': 'Q', 'LYS': 'K',
     'ILE': 'I', 'PRO': 'P', 'THR': 'T', 'PHE': 'F', 'ASN': 'N',
     'GLY': 'G', 'HIS': 'H', 'LEU': 'L', 'ARG': 'R', 'TRP': 'W',
     'ALA': 'A', 'VAL':'V', 'GLU': 'E', 'TYR': 'Y', 'MET': 'M', 'o': 'H', 'e': 'H', 'f': 'H',
                                       'H1S': 'H', 'H2S': 'H', 'H3S': 'H'}

    filtered_df = pd.read_csv('filtered_df_grouped.csv')

    # go through the files generated by foldx
    dir = '/Users/natalieso/Downloads/part_3_remote/'
    pdb_list = filtered_df['pdb_id'].unique()

    part_2_results = pd.read_csv('results/part2/part_2_result_splitted_for_part_3.csv', index_col=[0])
    error = []
    for pdb in pdb_list:

        filename = 'PS_'+pdb.lower()+'_scanning_output.txt'
        try:
            file1 = open(dir+filename, 'r')
        except:
            error.append(pdb)
        Lines = file1.readlines()

        for line in Lines:
            if not line:
                continue
            name, val = line.split('\t')
            original = name[:3]
            chain = name[3:4]
            target = name[-1]
            loc_str = ''
            name_with_origin = name[3:]
            for c in name_with_origin:
                if c.isdigit():
                    loc_str += c
            original_abbreviation = amino_acid_name_to_abbreviation[original]
            target = target if target not in amino_acid_name_to_abbreviation else amino_acid_name_to_abbreviation[target]

            if original_abbreviation == target:
                continue

            part_2_results.loc[(part_2_results['pdb_id'] == pdb) & (part_2_results['chain_id'] == chain)
                                              & (part_2_results['dssp_location'] == loc_str) &
                                              (part_2_results['amino_acid_original'] == original_abbreviation) &
                               (part_2_results['variants'] == target), 'DDG'] = val


        empty_ddg_rows = part_2_results.loc[pd.isnull(part_2_results['DDG']) | (part_2_results['DDG'] == ''), :]

        empty_ddg_rows.to_csv('empty_ddg_rows.csv')

        part_2_results.to_csv('part_3.csv')
    print('error', error)

if __name__ == "__main__":
    main()
